feat: Implement Mystery Teammate game components
- Add ConnectWalletButton for wallet connection functionality.
- Create GameFooter for navigation and game controls.
- Develop GameHeader to display user info and actions.
- Implement MysteryFlow to manage game state and teammate selection.
- Build MysteryTeammate component to encapsulate game logic and UI.
- Create UiShowcase for demonstrating UI components and interactions.
- Add UserProfileBadge for displaying user information with avatar.
